crawl/requirements.txt
beautifulsoup4==4.12.2
pyspark==3.3.1
pendulum
elasticsearch==8.6.0
numpy
pandas
pyarrow
pydantic
pyspark
pyyaml
google-cloud-storage
google-cloud-pubsub
pendulum